requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
pytest
